Update to new website versioning system
This new approach to versioning the website uses the workflow as the sole central orchestrator of the documentation
deployment system, restricting the role of the Python script (here renamed from "build.py" to "siteversion") to
solely to determining the appropriate version for the deployment. Since deployment of the website is the exclusive domain
of the automated systems, there is no point in using Task for that operation, so the deployment commands are moved from
the taskfile to the workflow.
