+++ docs/siteversion/tests/pytest.ini
@@ -1,0 +1,10 @@
+# Source: https://github.com/arduino/tooling-project-assets/blob/main/workflow-templates/assets/test-python/pytest.ini
+[pytest]
+filterwarnings =
+    error
+    ignore::DeprecationWarning
+    ignore::ResourceWarning
+
+# --capture=no - disable per-test capture
+# --tb=long sets the length of the traceback in case of failures
+addopts = --capture=no --tb=long --verbose
